refactor(react_hooks): split lib.rs into lib, scanner, helpers, and tests modules
Same shape as the earlier refactors. The 915-line crates/react_hooks/src/lib.rs
is split into four modules. All existing unit tests still pass.
